Remove monitor_id from segments and convert old data

+++ database/migrations/2019_06_14_211850_create_monitor_segment_table.php
@@ -15,8 +15,8 @@ class CreateMonitorSegmentTable extends Migration
     {
         Schema::create('monitor_segment', function (Blueprint $table) {
        		$table->increments('id');
-            $table->bigInteger('monitor_id');
-            $table->bigInteger('segment_id');
+            $table->unsignedInteger('monitor_id');
+            $table->unsignedInteger('segment_id');
             $table->timestamps();
 
             $table->foreign('monitor_id')->references('id')->on('monitors');
